Create getEmployees & getEmployee by ID controller & route. Update index.js for router connection.
