+++ routers/employees.js
@@ -1,8 +1,8 @@
 const express = require("express");
-const employeesController = require("../controllers/getEmployees");
+const employeesController = require("../controllers/employees");
 const router = express.Router();
 
 router.get("/", employeesController.getAllEmployees);
-router.get("/:id", employeesController.showEmployee);
+router.get("/:emp_no", employeesController.showEmployee);
 
 module.exports = router;
