--- routes/employees.js
@@ -1,5 +1,0 @@
-const express = require('express')
-const employeesController = require('../controllers/employees')
-const router = express.Router()
-
-module.exports = router
